// apps/raise/src/tender-user/user/types/user_status.ts
export type UserStatus =
  | 'WAITING_FOR_ACTIVATION'
  | 'SUSPENDED_ACCOUNT'
  | 'CANCELED_ACCOUNT'
  | 'ACTIVE_ACCOUNT'
  | 'REVISED_ACCOUNT'
  | 'WAITING_FOR_EDITING_APPROVAL';

export enum UserStatusEnum {
  WAITING_FOR_ACTIVATION = 'WAITING_FOR_ACTIVATION',
  SUSPENDED_ACCOUNT = 'SUSPENDED_ACCOUNT',
  CANCELED_ACCOUNT = 'CANCELED_ACCOUNT',
  ACTIVE_ACCOUNT = 'ACTIVE_ACCOUNT',
  REVISED_ACCOUNT = 'REVISED_ACCOUNT',
  WAITING_FOR_EDITING_APPROVAL = 'WAITING_FOR_EDITING_APPROVAL',
}